Outsource find functionality from View
