+++ rust/core-lib/src/view.rs
@@ -20,17 +20,14 @@ use std::ops::Range;
 use serde_json::Value;
 
 use xi_rope::rope::{Rope, LinesMetric, RopeInfo};
-use xi_rope::delta::{Delta, DeltaRegion};
+use xi_rope::delta::{Delta};
 use xi_rope::tree::{Cursor, Metric};
 use xi_rope::breaks::{Breaks, BreaksInfo, BreaksMetric, BreaksBaseMetric};
 use xi_rope::interval::Interval;
 use xi_rope::spans::Spans;
-use xi_rope::find::{find, CaseMatching};
 use xi_trace::trace_block;
-
 use client::Client;
 use edit_types::ViewEvent;
-use index_set::IndexSet;
 use line_cache_shadow::{self, LineCacheShadow, RenderPlan, RenderTactic};
 use movement::{Movement, region_movement, selection_movement};
 use rpc::{GestureType, MouseAction};
@@ -39,12 +36,11 @@ use selection::{Affinity, Selection, SelRegion};
 use tabs::{ViewId, BufferId};
 use width_cache::WidthCache;
 use word_boundaries::WordCursor;
-
+use find::Find;
 use linewrap;
 
 type StyleMap = RefCell<ThemeStyleMap>;
 
-const BACKWARDS_FIND_CHUNK_SIZE: usize = 32_768;
 
 /// A flag used to indicate when legacy actions should modify selections
 const FLAG_SELECT: u64 = 2;
@@ -73,20 +69,10 @@ pub struct View {
     /// description for the invariant.
     lc_shadow: LineCacheShadow,
 
-    /// The occurrences, which determine the highlights, have been updated.
-    hls_dirty: bool,
-
     /// New offset to be scrolled into position after an edit.
     scroll_to: Option<usize>,
 
-    /// The currently active search string
-    search_string: Option<String>,
-    /// The case matching setting for the currently active search
-    case_matching: CaseMatching,
-    /// The set of all known find occurrences (highlights)
-    occurrences: Option<Selection>,
-    /// Set of ranges that have already been searched for the currently active search string
-    valid_search: IndexSet,
+    find: Find,
 }
 
 /// The visual width of the buffer for the purpose of word wrapping.
@@ -133,11 +119,7 @@ impl View {
             breaks: None,
             wrap_col: WrapWidth::None,
             lc_shadow: LineCacheShadow::default(),
-            hls_dirty: true,
-            search_string: None,
-            case_matching: CaseMatching::CaseInsensitive,
-            occurrences: None,
-            valid_search: IndexSet::new(),
+            find: Find::new(),
         }
     }
 
@@ -163,7 +145,7 @@ impl View {
             Gesture { line, col, ty } =>
                 self.do_gesture(text, line, col, ty),
             GotoLine { line } => self.goto_line(text, line),
-            FindNext { wrap_around, allow_same } =>
+            FindNext { wrap_around, allow_same } =>         // todo
                 self.find_next(text, false,
                                wrap_around.unwrap_or(false),
                                allow_same.unwrap_or(false)),
@@ -213,7 +195,11 @@ impl View {
 
     fn do_cancel(&mut self, text: &Rope) {
         self.collapse_selections(text);
-        self.unset_find(text);
+        self.find.unset_find(text);
+    }
+
+    pub fn unset_find(&mut self, text: &Rope) {
+        self.find.unset_find(text)
     }
 
     fn goto_line(&mut self, text: &Rope, line: u64) {
@@ -483,9 +469,10 @@ impl View {
             }
         }
 
+        // todo
         let mut hls = Vec::new();
-        if let Some(ref occurrences) = self.occurrences {
-            for region in occurrences.regions_in_range(start_pos, pos) {
+        for search_occurrence in self.find.occurrences() {
+            for region in search_occurrence.regions_in_range(start_pos, pos) {
                 let sel_start_ix = clamp(region.min(), start_pos, pos) - start_pos;
                 let sel_end_ix = clamp(region.max(), start_pos, pos) - start_pos;
                 if sel_end_ix > sel_start_ix {
@@ -612,9 +599,11 @@ impl View {
                     } else {
                         let start_line = seg.our_line_num;
                         let end_line = start_line + seg.n;
-                        if self.hls_dirty {
+                        // todo
+                        if self.find.hls_dirty() {
                             self.update_find_for_lines(text, start_line, end_line);
                         }
+
                         let offset = self.offset_of_line(text, start_line);
                         let mut line_cursor = Cursor::new(text, offset);
                         let mut soft_breaks = self.breaks.as_ref().map(|breaks|
@@ -639,7 +628,7 @@ impl View {
         });
         client.update_view(self.view_id, &params);
         self.lc_shadow = b.build();
-        self.hls_dirty = false;
+        self.find.set_hls_dirty(false)
     }
 
     /// Update front-end with any changes to view since the last time sent.
@@ -785,31 +774,7 @@ impl View {
         // the front-end, but perhaps not for async edits.
         self.drag_state = None;
 
-        // Update search highlights for changed regions
-        if self.search_string.is_some() {
-            self.valid_search = self.valid_search.apply_delta(delta);
-            let mut occurrences = self.occurrences.take().unwrap_or_else(Selection::new);
-
-            // invalidate occurrences around deletion positions
-            for DeltaRegion{ old_offset, new_offset, len } in delta.iter_deletions() {
-                self.valid_search.delete_range(new_offset, new_offset + len);
-                occurrences.delete_range(old_offset, old_offset + len, false);
-            }
-
-            occurrences = occurrences.apply_delta(delta, false, false);
-
-            // invalidate occurrences around insert positions
-            for DeltaRegion{ new_offset, len, .. } in delta.iter_inserts() {
-                self.valid_search.delete_range(new_offset, new_offset + len);
-                occurrences.delete_range(new_offset, new_offset + len, false);
-            }
-
-            self.occurrences = Some(occurrences);
-
-            // update find for the whole delta (is going to only update invalid regions)
-            let (iv, _) = delta.summary();
-            self.update_find(text, iv.start(), iv.end(), true, false);
-        }
+        self.find.update_highlights(text, last_text, delta);      // todo
 
         // Note: for committing plugin edits, we probably want to know the priority
         // of the delta so we can set the cursor before or after the edit, as needed.
@@ -819,6 +784,7 @@ impl View {
 
     pub fn do_find(&mut self, text: &Rope, chars: Option<String>,
                    case_sensitive: bool) -> Value {
+        // todo
         let mut from_sel = false;
         let search_string = if chars.is_some() {
             chars
@@ -833,147 +799,15 @@ impl View {
             })
         };
 
-        if search_string.is_none() {
-            self.unset_find(text);
-            return Value::Null;
-        }
-
-        let search_string = search_string.unwrap();
-        if search_string.len() == 0 {
-            self.unset_find(text);
-            return Value::Null;
-        }
-
-        self.set_find(text, &search_string, case_sensitive);
-
-        Value::String(search_string.to_string())
-    }
-
-    /// Unsets the search and removes all highlights from the view.
-    pub(crate) fn unset_find(&mut self, text: &Rope) {
-        self.search_string = None;
-        self.occurrences = None;
-        self.hls_dirty = true;
-        // TODO: finer grained invalidation
-        self.set_dirty(text);
-        self.valid_search.clear();
-    }
-
-    /// Sets find for the view, highlights occurrences in the current viewport
-    /// and selects the first occurrence relative to the last cursor.
-    fn set_find(&mut self, text: &Rope, search_string: &str,
-                case_sensitive: bool) {
-        let case_matching = if case_sensitive {
-            CaseMatching::Exact
-        } else {
-            CaseMatching::CaseInsensitive
-        };
-
-        if let Some(ref s) = self.search_string {
-            if s == search_string && case_matching == self.case_matching {
-                // search parameters did not change
-                return;
-            }
-        }
-
-        self.unset_find(text);
-
-        self.search_string = Some(search_string.to_string());
-        self.case_matching = case_matching;
+        self.find.do_find(text, search_string, case_sensitive)
     }
 
     fn update_find_for_lines(&mut self, text: &Rope, first_line: usize, last_line: usize) {
-        if self.search_string.is_none() {
-            return;
-        }
         let start = self.offset_of_line(text, first_line);
         let end = self.offset_of_line(text, last_line);
-        self.update_find(text, start, end, true, false);
+        self.find.update_find(text, start, end, true, false);
     }
 
-    fn update_find(&mut self, text: &Rope, start: usize, end: usize, include_slop: bool,
-                   stop_on_found: bool)
-    {
-        if self.search_string.is_none() {
-            return;
-        }
-
-        let text_len = text.len();
-        // extend the search by twice the string length (twice, because case matching may increase
-        // the length of an occurrence)
-        let slop = if include_slop { self.search_string.as_ref().unwrap().len() * 2 } else { 0 };
-        let mut occurrences = self.occurrences.take().unwrap_or_else(Selection::new);
-        let mut searched_until = end;
-        let mut invalidate_from = None;
-
-        for (start, end) in self.valid_search.minus_one_range(start, end) {
-            let search_string = self.search_string.as_ref().unwrap();
-            let len = search_string.len();
-
-            // expand region to be able to find occurrences around the region's edges
-            let from = max(start, slop) - slop;
-            let to = min(end + slop, text.len());
-
-            // TODO: this interval might cut a unicode codepoint, make sure it is
-            // aligned to codepoint boundaries.
-            let text = text.subseq(Interval::new_closed_open(0, to));
-            let mut cursor = Cursor::new(&text, from);
-
-            while let Some(start) = find(&mut cursor, self.case_matching, &search_string) {
-                let end = start + len;
-
-                let region = SelRegion::new(start, end);
-                let prev_len = occurrences.len();
-                let (_, e) = occurrences.add_range_distinct(region);
-                // in case of ambiguous search results (e.g. search "aba" in "ababa"),
-                // the search result closer to the beginning of the file wins
-                if e != end {
-                    // Skip the search result and keep the occurrence that is closer to
-                    // the beginning of the file. Re-align the cursor to the kept
-                    // occurrence
-                    cursor.set(e);
-                    continue;
-                }
-
-                // add_range_distinct() above removes ambiguous regions after the added
-                // region, if something has been deleted, everything thereafter is
-                // invalidated
-                if occurrences.len() != prev_len + 1 {
-                    invalidate_from = Some(end);
-                    occurrences.delete_range(end, text_len, false);
-                    break;
-                }
-
-                if stop_on_found {
-                    searched_until = end;
-                    break;
-                }
-            }
-        }
-        self.occurrences = Some(occurrences);
-        if let Some(invalidate_from) = invalidate_from {
-            self.valid_search.union_one_range(start, invalidate_from);
-
-            // invalidate all search results from the point of the ambiguous search result until ...
-            let is_multi_line = LinesMetric::next(self.search_string.as_ref().unwrap(), 0).is_some();
-            if is_multi_line {
-                // ... the end of the file
-                self.valid_search.delete_range(invalidate_from, text_len);
-            } else {
-                // ... the end of the line
-                let mut cursor = Cursor::new(&text, invalidate_from);
-                if let Some(end_of_line) = cursor.next::<LinesMetric>() {
-                    self.valid_search.delete_range(invalidate_from, end_of_line);
-                }
-            }
-
-            // continue with the find for the current region
-            self.update_find(text, invalidate_from, end, false, false);
-        } else {
-            self.valid_search.union_one_range(start, searched_until);
-            self.hls_dirty = true;
-        }
-    }
 
     pub fn find_next(&mut self, text: &Rope, reverse: bool, wrap: bool, allow_same: bool) {
         self.select_next_occurrence(text, reverse, false, true, allow_same);
@@ -991,76 +825,12 @@ impl View {
     pub fn select_next_occurrence(&mut self, text: &Rope, reverse: bool, wrapped: bool,
                                   stop_on_found: bool, allow_same: bool)
     {
-        if self.search_string.is_none() {
-            return;
-        }
-
         let sel = match self.sel_regions().last() {
             Some(sel) => (sel.min(), sel.max()),
             None => return,
         };
 
-        let (from, to) = if reverse != wrapped { (0, sel.0) } else { (sel.0, text.len()) };
-        let mut next_occurrence;
-
-        loop {
-            next_occurrence = self.occurrences.as_ref().and_then(|occurrences| {
-                if occurrences.len() == 0 {
-                    return None;
-                }
-                if wrapped { // wrap around file boundaries
-                    if reverse {
-                        occurrences.last()
-                    } else {
-                        occurrences.first()
-                    }
-                } else {
-                    let ix = occurrences.search(sel.0);
-                    if reverse {
-                        ix.checked_sub(1).and_then(|i| occurrences.get(i))
-                    } else {
-                        occurrences.get(ix).and_then(|oc| {
-                            // if possible, the current selection should be extended, instead of
-                            // jumping to the next occurrence
-                            if oc.end == sel.1 && !allow_same {
-                                occurrences.get(ix+1)
-                            } else {
-                                Some(oc)
-                            }
-                        })
-                    }
-                }
-            }).cloned();
-
-            let region = {
-                let mut unsearched = self.valid_search.minus_one_range(from, to);
-                if reverse { unsearched.next_back() } else { unsearched.next() }
-            };
-            if let Some((b, e)) = region {
-                if let Some(ref occurrence) = next_occurrence {
-                    if (reverse && occurrence.start >= e) || (!reverse && occurrence.end <= b) {
-                        break;
-                    }
-                }
-
-                if !reverse {
-                    self.update_find(text, b, e, false, stop_on_found);
-                } else {
-                    // when searching backward, the actual search isn't executed backwards, which is
-                    // why the search is executed in chunks
-                    let start = if e - b > BACKWARDS_FIND_CHUNK_SIZE {
-                        e - BACKWARDS_FIND_CHUNK_SIZE
-                    } else {
-                        b
-                    };
-                    self.update_find(text, start, e, false, false);
-                }
-            } else {
-                break;
-            }
-        }
-
-        if let Some(occ) = next_occurrence {
+        if let Some(occ) = self.find.next_occurrence(text, reverse, wrapped, stop_on_found, allow_same, sel) {
             self.set_selection(text, occ);
         }
     }
